+++ forui/lib/src/widgets/popover/popover_controller.dart
@@ -18,16 +18,18 @@ class FPopoverController extends FChangeNotifier {
   late final AnimationController _animation;
   late final CurvedAnimation _curveScale;
   late final CurvedAnimation _curveFade;
-  late final Animation<double> _scale;
-  late final Animation<double> _fade;
+  late Animation<double> _scale;
+  late Animation<double> _fade;
 
-  /// Creates a [FPopoverController] with the given [vsync], [initial] and [motion].
-  FPopoverController({required TickerProvider vsync, double initial = 0.0, FPopoverMotion motion = const .new()}) {
+  /// Creates a [FPopoverController] with the given [vsync], [shown] and [motion].
+  FPopoverController({required TickerProvider vsync, bool shown = false, FPopoverMotion motion = const .new()}) {
+    shown ? _overlay.show() : _overlay.hide();
     _animation = AnimationController(
       vsync: vsync,
       duration: motion.entranceDuration,
       reverseDuration: motion.exitDuration,
-    )..value = initial;
+    )
+      ..value = shown ? 1 : 0;
     _curveFade = CurvedAnimation(parent: _animation, curve: motion.fadeInCurve, reverseCurve: motion.fadeOutCurve);
     _curveScale = CurvedAnimation(parent: _animation, curve: motion.expandCurve, reverseCurve: motion.collapseCurve);
     _scale = motion.scaleTween.animate(_curveScale);
@@ -81,37 +83,6 @@ class FPopoverController extends FChangeNotifier {
 
 @internal
 extension InternalFPopoverController on FPopoverController {
-  /// Updates the given [controller] based on the [shown] and [onChange].
-  ///
-  /// Typically used when updating a popover controller nested in another controller.
-  @useResult
-  static FPopoverController updateNested(
-    FPopoverController controller,
-    TickerProvider vsync,
-    bool? shown,
-    ValueChanged<bool>? onChange,
-  ) {
-    switch ((controller, shown != null)) {
-      // Lifted -> Lifted
-      case (final LiftedPopoverController lifted, true):
-        lifted.update(shown!, onChange!);
-        return lifted;
-
-      // Lifted -> Internal
-      case (LiftedPopoverController(), false):
-        controller.dispose();
-        return FPopoverController(vsync: vsync);
-
-      // Internal -> Lifted
-      case (_, true) when controller is! LiftedPopoverController:
-        controller.dispose();
-        return LiftedPopoverController(shown!, onChange!, vsync: vsync);
-
-      default:
-        return controller;
-    }
-  }
-
   OverlayPortalController get overlay => _overlay;
 
   Animation<double> get scale => _scale;
@@ -120,19 +91,32 @@ extension InternalFPopoverController on FPopoverController {
 }
 
 @internal
-class LiftedPopoverController extends FPopoverController {
+class ProxyPopoverController extends FPopoverController {
   int _monotonic;
-  void Function(bool shown) _onChange;
+  ValueChanged<bool> _onChange;
+  FPopoverMotion _motion;
 
-  LiftedPopoverController(bool shown, this._onChange, {required super.vsync, super.motion}) : _monotonic = 0 {
-    if (shown) {
-      _overlay.show();
-      _animation.value = 1;
-    }
-  }
+  ProxyPopoverController(this._onChange, this._motion, {required super.vsync, super.shown})
+      : _monotonic = 0,
+        super(motion: _motion);
 
-  void update(bool shown, void Function(bool shown) onChange) {
+  void update(bool shown, ValueChanged<bool> onChange, FPopoverMotion motion) {
     _onChange = onChange;
+    if (_motion != motion) {
+      _motion = motion;
+      _animation
+        ..duration = motion.entranceDuration
+        ..reverseDuration = motion.exitDuration;
+      _curveFade
+        ..curve = motion.fadeInCurve
+        ..reverseCurve = motion.fadeOutCurve;
+      _curveScale
+        ..curve = motion.expandCurve
+        ..reverseCurve = motion.collapseCurve;
+      _scale = motion.scaleTween.animate(_curveScale);
+      _fade = motion.fadeTween.animate(_curveFade);
+    }
+
     final current = ++_monotonic;
     SchedulerBinding.instance.addPostFrameCallback((_) async {
       if (current == _monotonic) {
@@ -160,13 +144,6 @@ class LiftedPopoverController extends FPopoverController {
 ///
 /// {@macro forui.foundation.doc_templates.control}
 sealed class FPopoverControl with Diagnosticable, _$FPopoverControlMixin {
-  /// Creates a [FPopoverControl].
-  const factory FPopoverControl.managed({
-    FPopoverController? controller,
-    FPopoverMotion? motion,
-    ValueChanged<bool>? onChange,
-  }) = FPopoverManagedControl;
-
   /// Creates a [FPopoverControl] for controlling a popover using lifted state.
   ///
   /// The [shown] parameter indicates whether the popover is currently shown.
@@ -177,14 +154,20 @@ sealed class FPopoverControl with Diagnosticable, _$FPopoverControlMixin {
     FPopoverMotion motion,
   }) = _Lifted;
 
+  /// Creates a [FPopoverControl].
+  const factory FPopoverControl.managed({
+    FPopoverController? controller,
+    bool? initial,
+    FPopoverMotion? motion,
+    ValueChanged<bool>? onChange,
+  }) = FPopoverManagedControl;
+
   const FPopoverControl._();
 
-  (FPopoverController, bool) _update(
-    FPopoverControl old,
-    FPopoverController controller,
-    VoidCallback callback,
-    TickerProvider vsync,
-  );
+  (FPopoverController, bool) _update(FPopoverControl old,
+      FPopoverController controller,
+      VoidCallback callback,
+      TickerProvider vsync,);
 }
 
 /// A [FPopoverManagedControl] enables widgets to manage their own controller internally while exposing parameters for
@@ -195,6 +178,13 @@ class FPopoverManagedControl extends FPopoverControl with Diagnosticable, _$FPop
   /// The controller.
   @override
   final FPopoverController? controller;
+
+  /// Whether the popover is initially shown. Defaults to false (hidden).
+  ///
+  /// ## Contract
+  /// Throws [AssertionError] if [initial] and [controller] are both provided.
+  @override
+  final bool? initial;
 
   /// The popover motion. Defaults to [FPopoverMotion].
   ///
@@ -208,13 +198,14 @@ class FPopoverManagedControl extends FPopoverControl with Diagnosticable, _$FPop
   final ValueChanged<bool>? onChange;
 
   /// Creates a [FPopoverControl].
-  const FPopoverManagedControl({this.controller, this.motion, this.onChange})
-    : assert(controller == null || motion == null, 'Cannot provide both controller and motion'),
-      super._();
+  const FPopoverManagedControl({this.controller, this.initial, this.motion, this.onChange})
+      : assert(controller == null || initial == null, 'Cannot provide both controller and initial.'),
+        assert(controller == null || motion == null, 'Cannot provide both controller and motion.'),
+        super._();
 
   @override
   FPopoverController createController(TickerProvider vsync) =>
-      controller ?? .new(vsync: vsync, motion: motion ?? const .new());
+      controller ?? .new(vsync: vsync, shown: initial ?? false, motion: motion ?? const .new());
 }
 
 class _Lifted extends FPopoverControl with _$_LiftedMixin {
@@ -229,9 +220,9 @@ class _Lifted extends FPopoverControl with _$_LiftedMixin {
 
   @override
   FPopoverController createController(TickerProvider vsync) =>
-      LiftedPopoverController(vsync: vsync, shown, onChange, motion: motion);
+      ProxyPopoverController(vsync: vsync, onChange, motion, shown: shown);
 
   @override
   void _updateController(FPopoverController controller, TickerProvider vsync) =>
-      (controller as LiftedPopoverController).update(shown, onChange);
+      (controller as ProxyPopoverController).update(shown, onChange, motion);
 }